Decider perf changes v2 (#69)
* Fix docsite description (for SEO)

* Introducing RequestHandler which encapsulates all HTTP operations. The JerseyRequestHandler is an implementation that uses the Jersey client.

* removed usages of Guava

* Typo fix (#3029)

* task summary ui (#3028)

* Add tests for updateTask

* Add callbackAfterSeconds and pollCount in the Task Summary tab

* additional constructors in JerseyRequestHandler to simplify its creation

* updated reminder about formatting

* Replace old logo asset with new logo

* Raise max results per page to 1000

* [UI] Separate workflowDefs from latestWorkflowDefs. Reuse existing react-query hooks wherever possible.

* [UI] Refactor task related data hooks.
* Replace workflow-by-task search with simple task search.
* Separate fetchers for poll data and queue size

* [UI] Support terminate with reason in Actions & Bulk Actions

* [UI] Search for Tasks directly

* [UI] Refactor Task Poll Data page to correctly account for domains.
* Add a Poll Data tab to task panel

* [UI] Fix duration formatting

* [UI] Update views to use refactored workflow data hooks and Dropdowns

* [UI] Refactor Do-while loops

* [UI] Use new logo in AppLogo

* improve performance of checkForWorkflowCompletion

* sdk documentation additions (#3047)

* sdk documentation additions

* csharp update

* Bugfix/jointask output (#3055)

* re-throw TerminateWorkflowException for task update

* Add only non-empty task output for Join

* decider changes

* fixes

* Update build.gradle

* updates

* terminate

* Update ForkJoinDynamicTaskMapper.java

* Update SwitchTaskMapper.java

* more changes

* disable tests (temp)

* formatting

+++ core/src/main/java/com/netflix/conductor/core/reconciliation/WorkflowSweeper.java
@@ -75,7 +75,7 @@ public class WorkflowSweeper {
             }
 
             WorkflowModel workflow = workflowExecutor.decide(workflowId);
-            if (workflow.getStatus().isTerminal()) {
+            if (workflow != null && workflow.getStatus().isTerminal()) {
                 queueDAO.remove(DECIDER_QUEUE, workflowId);
                 return;
             }
